+++ HISTORY.md
@@ -45,6 +45,9 @@
   * Add aliases for previous names for backwards-compatibility
 
 -----
+### 0.7.5 (2021-TBD)
+* Fix incorrect location of `redirects.sqlite` when using filesystem backend
+
 ### 0.7.4 (2021-08-16)
 * Fix an issue with httpdate strings from `Expires` headers not getting converted to UTC
 * Fix a packaging issue with extra files added to top-level wheel directory